feat: extracted cache features to its dedicated class
