+++ providers/keycloak/src/airflow/providers/keycloak/auth_manager/constants.py
@@ -25,4 +25,3 @@ CONF_REALM_KEY = "realm"
 CONF_SERVER_URL_KEY = "server_url"
 CONF_DAG_PERMISSIONS_CACHE_TTL_KEY = "dag_permissions_cache_ttl"
 CONF_DAG_INVENTORY_CACHE_TTL_KEY = "dag_inventory_cache_ttl"
-CONF_AUTHORIZATION_PARALLELISM_KEY = "authorization_parallelism"
