+++ providers/keycloak/src/airflow/providers/keycloak/auth_manager/keycloak_auth_manager.py
@@ -19,19 +19,12 @@ from __future__ import annotations
 import argparse
 import json
 import logging
-import threading
-import weakref
-from collections import defaultdict
-from collections.abc import Iterable
-from concurrent.futures import ThreadPoolExecutor, as_completed
-from time import monotonic
-from typing import TYPE_CHECKING, Any, NamedTuple, cast
+from typing import TYPE_CHECKING, Any, cast
 from urllib.parse import urljoin
 
 import requests
 from fastapi import FastAPI
 from keycloak import KeycloakOpenID
-from sqlalchemy import select
 
 from airflow.api_fastapi.app import AUTH_MANAGER_FASTAPI_APP_PREFIX
 from airflow.api_fastapi.auth.managers.base_auth_manager import BaseAuthManager
@@ -45,15 +38,15 @@ from airflow.api_fastapi.common.types import MenuItem
 from airflow.cli.cli_config import CLICommand, DefaultHelpParser, GroupCommand
 from airflow.configuration import conf
 from airflow.exceptions import AirflowException
-from airflow.models import DagModel
-from airflow.models.dagbundle import DagBundleModel
+from airflow.providers.keycloak.auth_manager.cache import (
+    ContextAttributes,
+    KeycloakDagPermissionCache,
+    OptionalContextAttributes,
+)
 from airflow.providers.keycloak.auth_manager.cli.definition import KEYCLOAK_AUTH_MANAGER_COMMANDS
 from airflow.providers.keycloak.auth_manager.constants import (
-    CONF_AUTHORIZATION_PARALLELISM_KEY,
     CONF_CLIENT_ID_KEY,
     CONF_CLIENT_SECRET_KEY,
-    CONF_DAG_INVENTORY_CACHE_TTL_KEY,
-    CONF_DAG_PERMISSIONS_CACHE_TTL_KEY,
     CONF_REALM_KEY,
     CONF_SECTION_NAME,
     CONF_SERVER_URL_KEY,
@@ -61,11 +54,8 @@ from airflow.providers.keycloak.auth_manager.constants import (
 from airflow.providers.keycloak.auth_manager.resources import KeycloakResource
 from airflow.providers.keycloak.auth_manager.user import KeycloakAuthManagerUser
 from airflow.utils.helpers import prune_dict
-from airflow.utils.session import NEW_SESSION, provide_session
 
 if TYPE_CHECKING:
-    from sqlalchemy.orm import Session
-
     from airflow.api_fastapi.auth.managers.base_auth_manager import ResourceMethod
     from airflow.api_fastapi.auth.managers.models.resource_details import (
         AccessView,
@@ -83,14 +73,6 @@ if TYPE_CHECKING:
 log = logging.getLogger(__name__)
 
 RESOURCE_ID_ATTRIBUTE_NAME = "resource_id"
-ContextAttributes = dict[str, str | None]
-OptionalContextAttributes = ContextAttributes | None
-CacheKey = tuple[str, str, ExtendedResourceMethod, str | None]
-
-
-class _DagPermissionCacheEntry(NamedTuple):
-    expires_at: float
-    allowed: bool
 
 
 def get_parser() -> argparse.ArgumentParser:
@@ -113,41 +95,13 @@ class KeycloakAuthManager(BaseAuthManager[KeycloakAuthManagerUser]):
 
     def __init__(self) -> None:
         super().__init__()
-        self._dag_permissions_cache_ttl_seconds = conf.getint(
-            CONF_SECTION_NAME,
-            CONF_DAG_PERMISSIONS_CACHE_TTL_KEY,
-            fallback=30,
+        self._dag_cache = KeycloakDagPermissionCache(
+            permission_resolver=self._authorize_dag,
         )
-        self._dag_inventory_cache_ttl_seconds = conf.getint(
-            CONF_SECTION_NAME,
-            CONF_DAG_INVENTORY_CACHE_TTL_KEY,
-            fallback=300,
-        )
-        self._dag_permissions_cache: dict[CacheKey, _DagPermissionCacheEntry] = {}
-        self._dag_permissions_cache_lock = threading.RLock()
-        self._dag_inventory: dict[str | None, set[str]] | None = None
-        self._dag_inventory_last_refreshed: float | None = None
-        self._dag_inventory_lock = threading.RLock()
-        self._authorization_parallelism = max(
-            1,
-            conf.getint(
-                CONF_SECTION_NAME,
-                CONF_AUTHORIZATION_PARALLELISM_KEY,
-                fallback=8,
-            ),
-        )
-        self._dag_warmup_executor = ThreadPoolExecutor(max_workers=2, thread_name_prefix="keycloak-dag-warm")
-        self._warmup_executor_finalizer = weakref.finalize(
-            self, self._dag_warmup_executor.shutdown, wait=False
-        )
-        self._dag_warmup_lock = threading.RLock()
-        self._dag_warmup_inflight: set[str] = set()
 
     def init(self) -> None:
         super().init()
-        if self._dag_permissions_cache_ttl_seconds <= 0:
-            return
-        self._dag_warmup_executor.submit(self._prime_dag_inventory)
+        self._dag_cache.init()
 
     def deserialize_user(self, token: dict[str, Any]) -> KeycloakAuthManagerUser:
         return KeycloakAuthManagerUser(
@@ -222,153 +176,14 @@ class KeycloakAuthManager(BaseAuthManager[KeycloakAuthManagerUser]):
             attributes=attributes,
         )
 
-    def _get_cached_dag_permission(
-        self,
-        cache_key: CacheKey,
-        *,
-        now: float,
-    ) -> bool | None:
-        if self._dag_permissions_cache_ttl_seconds <= 0:
-            return None
-
-        with self._dag_permissions_cache_lock:
-            cached = self._dag_permissions_cache.get(cache_key)
-            if not cached:
-                return None
-            if cached.expires_at <= now:
-                self._dag_permissions_cache.pop(cache_key, None)
-                return None
-            return cached.allowed
-
-    def _set_cached_dag_permission(
-        self,
-        cache_key: CacheKey,
-        allowed: bool,
-        *,
-        now: float,
-    ) -> None:
-        if self._dag_permissions_cache_ttl_seconds <= 0:
-            return
-
-        expires_at = now + self._dag_permissions_cache_ttl_seconds
-        with self._dag_permissions_cache_lock:
-            self._dag_permissions_cache[cache_key] = _DagPermissionCacheEntry(expires_at, allowed)
-
-    def _prime_dag_inventory(self) -> None:
-        try:
-            inventory = self._fetch_dag_inventory()
-            with self._dag_inventory_lock:
-                self._dag_inventory = inventory
-                self._dag_inventory_last_refreshed = monotonic()
-        except Exception:
-            log.exception("Failed to prime DAG inventory for permissions warmup")
-
-    def _get_dag_inventory(self) -> dict[str | None, set[str]]:
-        if self._dag_inventory_cache_ttl_seconds <= 0:
-            return self._fetch_dag_inventory()
-
-        with self._dag_inventory_lock:
-            inventory = self._dag_inventory
-            last_refreshed = self._dag_inventory_last_refreshed
-
-        current_time = monotonic()
-        if (
-            inventory is not None
-            and last_refreshed is not None
-            and current_time - last_refreshed < self._dag_inventory_cache_ttl_seconds
-        ):
-            return {team: set(dag_ids) for team, dag_ids in inventory.items()}
-
-        inventory = self._fetch_dag_inventory()
-        with self._dag_inventory_lock:
-            self._dag_inventory = inventory
-            self._dag_inventory_last_refreshed = monotonic()
-        return inventory
-
-    @staticmethod
-    @provide_session
-    def _fetch_dag_inventory(*, session: Session = NEW_SESSION) -> dict[str | None, set[str]]:
-        team_model_cls: type[Any] | None
-        team_assoc_table: Any | None
-        try:
-            # Lazy import: Team models exist only in newer Airflow releases
-            from airflow.models.team import (
-                Team as team_model_cls_runtime,
-                dag_bundle_team_association_table as team_assoc_table_runtime,
-            )
-        except ModuleNotFoundError:
-            team_model_cls = None
-            team_assoc_table = None
-        else:
-            team_model_cls = team_model_cls_runtime
-            team_assoc_table = team_assoc_table_runtime
-
-        if team_model_cls is None or team_assoc_table is None:
-            stmt = select(DagModel.dag_id)
-            dag_ids = session.execute(stmt).scalars().all()
-            return {None: set(dag_ids)}
-
-        stmt = (
-            select(DagModel.dag_id, team_model_cls.name)
-            .join(DagBundleModel, DagModel.bundle_name == DagBundleModel.name)
-            .join(
-                team_assoc_table,
-                DagBundleModel.name == team_assoc_table.c.dag_bundle_name,
-                isouter=True,
-            )
-            .join(
-                team_model_cls,
-                team_model_cls.id == team_assoc_table.c.team_id,
-                isouter=True,
-            )
-        )
-        rows = session.execute(stmt).all()
-        dags_by_team: dict[str | None, set[str]] = defaultdict(set)
-        for dag_id, team_name in rows:
-            dags_by_team[team_name].add(dag_id)
-        return {team: set(dag_ids) for team, dag_ids in dags_by_team.items()}
-
-    def _warmup_user_dag_permissions(
-        self,
-        user: KeycloakAuthManagerUser,
-        method: ResourceMethod = "GET",
-    ) -> None:
-        user_id = str(user.get_id())
-        method_value = cast("ExtendedResourceMethod", method)
-        try:
-            dag_inventory = self._get_dag_inventory()
-            if not dag_inventory:
-                return
-
-            for team_name, dag_ids in dag_inventory.items():
-                if not dag_ids:
-                    continue
-                attributes: OptionalContextAttributes = {"team_name": team_name} if team_name else None
-
-                results = self._check_dag_authorizations(
-                    dag_ids,
-                    user=user,
-                    method=method_value,
-                    attributes=attributes,
-                    log_context={"team": team_name},
-                )
-                cache_time = monotonic()
-                for dag_id, allowed in results.items():
-                    cache_key = (user_id, dag_id, method_value, team_name)
-                    self._set_cached_dag_permission(cache_key, allowed, now=cache_time)
-        except Exception:
-            log.exception("Failed to warm DAG permissions for user %s", user_id)
-        finally:
-            with self._dag_warmup_lock:
-                self._dag_warmup_inflight.discard(user_id)
-
     def schedule_dag_permission_warmup(
         self,
         user: KeycloakAuthManagerUser,
         *,
         method: ResourceMethod = "GET",
     ) -> None:
-        if self._dag_permissions_cache_ttl_seconds <= 0:
+        method_value = cast("ExtendedResourceMethod", method)
+        if not self._dag_cache.permissions_cache_enabled:
             return
 
         user_snapshot = KeycloakAuthManagerUser(
@@ -377,12 +192,10 @@ class KeycloakAuthManager(BaseAuthManager[KeycloakAuthManagerUser]):
             access_token=getattr(user, "access_token", ""),
             refresh_token=getattr(user, "refresh_token", ""),
         )
-        user_id = user_snapshot.get_id()
-        with self._dag_warmup_lock:
-            if user_id in self._dag_warmup_inflight:
-                return
-            self._dag_warmup_inflight.add(user_id)
-        self._dag_warmup_executor.submit(self._warmup_user_dag_permissions, user_snapshot, method)
+        self._dag_cache.schedule_dag_permission_warmup(
+            user_snapshot,
+            method=method_value,
+        )
 
     def filter_authorized_dag_ids(
         self,
@@ -392,44 +205,13 @@ class KeycloakAuthManager(BaseAuthManager[KeycloakAuthManagerUser]):
         method: ResourceMethod = "GET",
         team_name: str | None = None,
     ) -> set[str]:
-        if not dag_ids:
-            return set()
-
-        user_id = str(user.get_id())
         method_value = cast("ExtendedResourceMethod", method)
-        lookup_attributes: OptionalContextAttributes = {"team_name": team_name} if team_name else None
-
-        now = monotonic()
-        authorized_dags: set[str] = set()
-        dag_ids_to_check: list[tuple[str, CacheKey]] = []
-        for dag_id in dag_ids:
-            cache_key = (user_id, dag_id, method_value, team_name)
-            cached_permission = self._get_cached_dag_permission(cache_key, now=now)
-            if cached_permission is None:
-                dag_ids_to_check.append((dag_id, cache_key))
-                continue
-            if cached_permission:
-                authorized_dags.add(dag_id)
-
-        if dag_ids_to_check:
-            dag_ids_list = [dag_id for dag_id, _ in dag_ids_to_check]
-            results = self._check_dag_authorizations(
-                dag_ids_list,
-                user=user,
-                method=method_value,
-                attributes=lookup_attributes,
-                log_context={"user_id": user_id},
-            )
-            cache_time = monotonic()
-            for dag_id, cache_key in dag_ids_to_check:
-                is_authorized = results.get(dag_id)
-                if is_authorized is None:
-                    continue
-                if is_authorized:
-                    authorized_dags.add(dag_id)
-                self._set_cached_dag_permission(cache_key, is_authorized, now=cache_time)
-
-        return authorized_dags
+        return self._dag_cache.filter_authorized_dag_ids(
+            dag_ids=dag_ids,
+            user=user,
+            method=method_value,
+            team_name=team_name,
+        )
 
     def is_authorized_backfill(
         self, *, method: ResourceMethod, user: KeycloakAuthManagerUser, details: BackfillDetails | None = None
@@ -470,52 +252,20 @@ class KeycloakAuthManager(BaseAuthManager[KeycloakAuthManagerUser]):
             method=method, resource_type=KeycloakResource.VARIABLE, user=user, resource_id=variable_key
         )
 
-    def _check_dag_authorizations(
+    def _authorize_dag(
         self,
-        dag_ids: Iterable[str],
-        *,
-        user: KeycloakAuthManagerUser,
         method: ExtendedResourceMethod,
+        user: KeycloakAuthManagerUser,
+        dag_id: str,
         attributes: OptionalContextAttributes = None,
-        log_context: OptionalContextAttributes = None,
-    ) -> dict[str, bool]:
-        dag_list = list(dag_ids)
-        if not dag_list:
-            return {}
-
-        max_workers = min(self._authorization_parallelism, len(dag_list))
-        if max_workers <= 0:
-            max_workers = 1
-
-        user_id = str(user.get_id())
-        context_suffix = ""
-        if log_context:
-            decorated = [f"{key}={value}" for key, value in log_context.items() if value is not None]
-            if decorated:
-                context_suffix = f" ({', '.join(decorated)})"
-
-        results: dict[str, bool] = {}
-        with ThreadPoolExecutor(max_workers=max_workers) as executor:
-            futures = {
-                executor.submit(
-                    self._is_authorized,
-                    method=method,
-                    resource_type=KeycloakResource.DAG,
-                    user=user,
-                    resource_id=dag_id,
-                    attributes=attributes,
-                ): dag_id
-                for dag_id in dag_list
-            }
-
-            for future in as_completed(futures):
-                dag_id = futures[future]
-                try:
-                    results[dag_id] = future.result()
-                except Exception:
-                    log.exception("Failed to authorize dag %s for user %s%s", dag_id, user_id, context_suffix)
-
-        return results
+    ) -> bool:
+        return self._is_authorized(
+            method=method,
+            resource_type=KeycloakResource.DAG,
+            user=user,
+            resource_id=dag_id,
+            attributes=attributes,
+        )
 
     def is_authorized_pool(
         self, *, method: ResourceMethod, user: KeycloakAuthManagerUser, details: PoolDetails | None = None
